User-stacks are no longer allocated by default

+++ include/kernel/mm/vm.h
@@ -90,7 +90,7 @@ void vm_unmap_vaddr(uintptr_t virt_addr);
 
 page_t *vm_get_page(uintptr_t addr);
 vm_area_t *vm_get_vm_area(uintptr_t addr);
-int vm_update_attributes(uintptr_t addr, uint8_t attr, void *ptr);
+int vm_update_attributes(uintptr_t addr, uint32_t attr, void *ptr);
 int vm_attach_page(uintptr_t addr, page_t *page);
 int vm_attach_file(uintptr_t addr, file_t *file);
 bool vm_find_free_area(vm_search_t search_type, uintptr_t *addr, size_t len);
